requirements.txt
fastapi~=0.65.2
SQLAlchemy~=1.4.18
pydantic~=1.8.2
python-dotenv~=0.17.1
uvicorn~=0.14.0
sqlalchemy-cockroachdb
psycopg2
pyodbc
Pillow~=8.2.0
fastapi-jwt-auth
azure-cognitiveservices-vision-face
google-cloud-storage
azure-storage-blob==12.11.0
requests~=2.25.1
msrest~=0.6.21
azure-cognitiveservices-vision-customvision
